+++ functions/env.d.ts
@@ -4,6 +4,12 @@
 
 interface Env {
   UMAMI_BACKEND_URL: string
+  USER_BLACKLIST: KVNamespace
+}
+
+interface KVNamespace {
+  get: (key: string, options?: { type?: 'text' }) => Promise<string | null>
+  put: (key: string, value: string, options?: { expirationTtl?: number }) => Promise<void>
 }
 
 interface EventContext<Environment = Env> {